+++ src/main/java/fr/maxlego08/menu/ZInventoryManager.java
@@ -352,6 +352,7 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
         // Load actions
         buttonManager.registerAction(new BroadcastLoader(this.plugin));
         buttonManager.registerAction(new MessageLoader());
+        buttonManager.registerAction(new MessageToLoader());
         buttonManager.registerAction(new BookLoader());
         buttonManager.registerAction(new SoundLoader());
         buttonManager.registerAction(new BroadcastSoundLoader());